+++ HISTORY.md
@@ -2,6 +2,7 @@
 
 ## 1.1.0-pre ##
 
+* Added test for invalid width and height properties ([#3](https://github.com/AurelioDeRosa/Audero-Smoke-Effect/issues/3))
 * Created `HISTORY.md` to trace changes ([#4](https://github.com/AurelioDeRosa/Audero-Smoke-Effect/issues/4))
 * Fixed the problem of resources not released after repeat number is reached ([#2](https://github.com/AurelioDeRosa/Audero-Smoke-Effect/issues/2))
 * Fixed the problem of elements selected in a call that share the same settings object ([#1](https://github.com/AurelioDeRosa/Audero-Smoke-Effect/issues/1))